+++ cmd/keytransparency-server/main.go
@@ -46,8 +46,8 @@ import (
 
 	cmutation "github.com/google/keytransparency/core/mutation"
 	pb "github.com/google/keytransparency/core/proto/keytransparency_v1"
+	mpb "github.com/google/keytransparency/core/proto/mutation_v1_service"
 	gauth "github.com/google/keytransparency/impl/google/authentication"
-	mpb "github.com/google/keytransparency/impl/proto/mutation_v1_service"
 	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
 )
