+++ impl/monitor/client/client.go
@@ -27,7 +27,7 @@ import (
 	"google.golang.org/grpc"
 
 	ktpb "github.com/google/keytransparency/core/proto/keytransparency_v1"
-	mupb "github.com/google/keytransparency/impl/proto/mutation_v1_service"
+	mupb "github.com/google/keytransparency/core/proto/mutation_v1_service"
 )
 
 // Each page contains pageSize profiles. Each profile contains multiple
